OrderService/requirements.txt
anyio==3.3.4
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
decorator==5.1.0
dependency-injector==4.36.2
fastapi==0.70.0
h11==0.12.0
idna==3.3
pika==1.2.0
py==1.10.0
pydantic==1.8.2
retry==0.9.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
